restructured to make more sense
