+++ ai/tools/README.md
@@ -1,12 +1,12 @@
 # Tool System Architecture
 
-The tool system provides a modular way to add capabilities to the AI agent. Tools are registered based on configuration permissions and can be easily extended.
+The tool system provides a modular way to add capabilities to the AI agent. Tools are self-contained, registered automatically based on permissions, and easy to extend.
 
 ## Structure
 
 ### Core Components
 
-- **`tools.go`**: Defines the `Tool` interface and `Registry` for managing tools
+- **`tools.go`**: Defines the `Tool` interface, `Registry`, and automatic tool registration
 - **`file_tools.go`**: Implements file operation tools (read, write, list directory)
 
 ### Tool Interface
@@ -22,12 +22,14 @@ type Tool interface {
 }
 ```
 
-### Permission-Based Registration
+### Registration
 
-Tools are automatically filtered based on config permissions:
+Tools are automatically registered in `tools.NewRegistry()` and filtered based on config permissions:
 - `Read`: Enables `read_file` and `list_directory`
 - `Write`: Enables `write_file`
 - `Command`: Reserved for shell/exec tools (future)
+
+The registry is created once and passed to the chat manager, ensuring clean separation of concerns.
 
 ## Available Tools
 
@@ -48,12 +50,42 @@ Tools are automatically filtered based on config permissions:
 
 ## Adding New Tools
 
-1. Create a new struct implementing the `Tool` interface
-2. Register it in `fileops.NewManager()`:
+1. Create a new struct implementing the `Tool` interface in the `tools/` package
+2. Register it in `tools.NewRegistry()`:
    ```go
-   toolRegistry.Register(NewYourTool(cfg))
+   r.Register(NewYourTool(cfg))
    ```
 3. The registry handles permission checking automatically
+
+**Example:**
+```go
+// In tools/your_tool.go
+type YourTool struct {
+    cfg *config.Config
+}
+
+func NewYourTool(cfg *config.Config) *YourTool {
+    return &YourTool{cfg: cfg}
+}
+
+// Implement Tool interface methods...
+
+// In tools/tools.go NewRegistry() function:
+r.Register(NewYourTool(cfg))
+```
+
+## Architecture
+
+```
+server/
+  └─ Creates toolRegistry and chatManager
+         ↓
+      tools/
+        └─ Self-contained tool registration
+               ↓
+            chat/
+              └─ Uses tool registry for LLM interactions
+```
 
 ## Usage
